chore: split tyewriter transition to another file
