+++ src/routes/+page.svelte
@@ -1,4 +1,5 @@
 <script>
+	import { typewriter } from '$lib';
 	import { onMount } from 'svelte';
 
 	let day;
@@ -8,25 +9,6 @@
 	onMount(() => {
 		day = new Date().getDay();
 	});
-
-	function typewriter(node, { speed = 1 }) {
-		const valid = node.childNodes.length === 1 && node.childNodes[0].nodeType === Node.TEXT_NODE;
-
-		if (!valid) {
-			throw new Error(`This transition only works on elements with a single text node child`);
-		}
-
-		const text = node.textContent;
-		const duration = text.length / (speed * 0.01);
-
-		return {
-			duration,
-			tick: (t) => {
-				const i = Math.trunc(text.length * t);
-				node.textContent = text.slice(0, i);
-			}
-		};
-	}
 </script>
 
 <div class="flex-grow flex flex-col items-center justify-center">
